+++ app/src/main/java/com/stopkaaaa/androidacademyproject/ui/MovieClickListener.kt
@@ -1,6 +1,6 @@
 package com.stopkaaaa.androidacademyproject.ui
 
 interface MovieClickListener {
-    fun movieClicked(movieId: Long)
+    fun movieClicked(movieId: Int)
     fun backPressed()
 }
